feat: Add types-pyyaml dependency and refactor server info handling
- Introduced `types-pyyaml` as a development dependency in `pyproject.toml` and `uv.lock`.
- Refactored the `ServerInfo` class to utilize shared types from the `common` module, enhancing code clarity and maintainability.
- Updated the `BaseServer` class to reflect changes in `ServerInfo` and improved health status management.
- Adjusted imports in various modules to streamline access to common types.
