+++ src/common/__init__.py
@@ -1,0 +1,21 @@
+"""Common types and utilities shared across lightfast-mcp core and tools."""
+
+from .types import (
+    HealthStatus,
+    OperationStatus,
+    ServerInfo,
+    ServerState,
+    ToolCall,
+    ToolCallState,
+    ToolResult,
+)
+
+__all__ = [
+    "HealthStatus",
+    "OperationStatus",
+    "ServerInfo",
+    "ServerState",
+    "ToolCall",
+    "ToolCallState",
+    "ToolResult",
+]
