+++ src/lightfast_mcp/core/__init__.py
@@ -1,9 +1,18 @@
 """Core infrastructure for MCP server implementations."""
 
-from .base_server import BaseServer, ServerConfig, ServerInfo
+# Import shared types from common module
+import sys
+from pathlib import Path
+
+from .base_server import BaseServer, ServerConfig
+
+sys.path.insert(0, str(Path(__file__).parent.parent.parent))
+from common import HealthStatus, ServerInfo, ServerState
 
 __all__ = [
     "BaseServer",
     "ServerConfig",
     "ServerInfo",
+    "ServerState",
+    "HealthStatus",
 ]
